+++ src/pages/Blog.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import { fetchBlogPosts } from '../api/blogService';
+import { fetchBlogPosts } from '../services/blogService';
 import ROUTES from '../routes';
 
 interface BlogPost {
