+++ src/services/axiosInstance.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 const API_BASE_URL =
   process.env.NODE_ENV === 'development'
-    ? '/'
+    ? import.meta.env.VITE_RUBRION_DEV_URL
     : import.meta.env.VITE_RUBRION_API_URL;
 
 if (!API_BASE_URL) {
